Employ same Div for selection and showing user info
