+++ user-interface/frontend/themes/datamanager/components/dialog.css
@@ -128,41 +128,6 @@ Since we want to remove the spacing between the cancel and confirm button we rep
   font-weight: bold;
 }
 
-.collaborator-info {
-  display: flex;
-  align-items: center;
-  gap: var(--lumo-space-l);
-}
-
-.collaborator-info .avatar {
-  width: var(--lumo-icon-size-l);
-  height: var(--lumo-icon-size-l);
-}
-
-.collaborator-info .user-info {
-  display: flex;
-  gap: var(--lumo-space-m);
-  flex-direction: row;
-  flex-wrap: wrap;
-}
-
-.collaborator-info .user-info .oidc {
-  display: inline-flex;
-  align-items: center;
-  gap: var(--lumo-space-s);
-}
-
-.collaborator-info .user-info .info-item {
-  display: flex;
-  flex-direction: column;
-  gap: var(--lumo-space-s);
-}
-
-.collaborator-info .user-info .info-item .label {
-  font-size: var(--lumo-font-size-s);
-  color: var(--lumo-secondary-text-color)
-}
-
 .batch-registration-dialog::part(content),
 .batch-update-dialog::part(content) {
   display: flex;
