+++ user-interface/frontend/themes/datamanager/components/div.css
@@ -4,31 +4,32 @@
   flex-flow: row wrap;
 }
 
-.user-info-cell-component {
+.user-info-component {
   display: flex;
   align-items: center;
   gap: var(--lumo-space-l);
 }
 
-.user-info-cell-component .avatar {
+.user-info-component .avatar {
   width: var(--lumo-icon-size-l);
   height: var(--lumo-icon-size-l);
 }
 
-.user-info-cell-component .user-info {
+.user-info-component .user-info {
   display: flex;
   flex-direction: column;
   flex-wrap: wrap;
   gap: var(--lumo-space-s);
 }
 
-.user-info-cell-component .user-info .oidc {
+.user-info-component .user-info .oidc {
   display: inline-flex;
   align-items: center;
   gap: var(--lumo-space-s);
+  white-space: nowrap;
 }
 
-.user-info-cell-component .user-info .user-name-and-full-name {
+.user-info-component .user-info .user-name-and-full-name {
   display: flex;
   gap: var(--lumo-space-s);
   align-items: baseline;
